requirements.txt
selenium>=4.15.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0
openai>=1.0.0
anthropic>=0.5.0
